% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{as.character.RsyncD}
\alias{as.character.RsyncD}
\title{Rsync API}
\usage{
\method{as.character}{RsyncD}(x, ...)
}
\arguments{
\item{x}{Rsync object}

\item{...}{more arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
converts a RsynsServer object to a character
}
}
