% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{as.character.RsyncDHTTP}
\alias{as.character.RsyncDHTTP}
\title{Rsync API}
\usage{
\method{as.character}{RsyncDHTTP}(x, ...)
}
\arguments{
\item{x}{RsyncServer object}

\item{...}{more arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
converts a RsynsServer object to a character
}
}
