% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awss3.R, R/getFile.R, R/listFiles.R,
%   R/removeFile.R, R/sendFile.R
\name{awss3}
\alias{awss3}
\alias{getFile.awss3}
\alias{listFiles.awss3}
\alias{removeFile.awss3}
\alias{sendFile.awss3}
\title{Connection object to a AWS S3 bucket}
\usage{
awss3(dest, src = getwd(), profile = NULL)

\method{getFile}{awss3}(db, fileName, validate = FALSE,
  verbose = FALSE, ...)

\method{listFiles}{awss3}(db, ...)

\method{removeFile}{awss3}(db, fileName, verbose = FALSE, ...)

\method{sendFile}{awss3}(db, fileName, validate = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{dest}{(character) the address to the rsync daemon or a folder.}

\item{src}{(character) a folder.}

\item{db}{(rsync) an object of class 'rsync' initialized with \code{rsync}.}

\item{fileName}{(character) a file name that exists in \code{src}}

\item{validate}{(logical) if the file in dest and src should be validated
using a sha256 check sum.}

\item{verbose}{(logical) if we use 'vorbose' as option in the cli.}

\item{...}{arguments passed to methods.}
}
\description{
Connection object to a AWS S3 bucket
}
