% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsynccli.R
\name{checkSystemResult}
\alias{checkSystemResult}
\title{Check System Command Result for Exit Code}
\usage{
checkSystemResult(result)
}
\arguments{
\item{result}{The result of a \code{system()} function execution. This can be:
\itemize{
  \item A character vector containing the command's output.
  \item A numeric value representing the exit status code.
  \item An object with a \code{status} attribute.
}}
}
\value{
\itemize{
    \item If \code{result} is a character vector (command output) and has no
      \code{status} attribute, returns the output.
    \item If \code{result} is numeric (exit status code), returns the status
      code.
    \item If the command failed (non-zero exit status), the function stops
      with an error message.
  }
}
\description{
Evaluates the result from R's \code{system()} function and checks for a
non-zero exit status. If the system command failed (i.e., returned a non-zero
exit status), the function throws an error. If the result contains the
command's output, it is returned.
}
