% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{delete}
\alias{delete}
\title{Rsync API}
\usage{
delete(db, entryName, verbose = FALSE)
}
\arguments{
\item{db}{RsyncServer object , returning: 'name', 'lastModified' and 'size'}

\item{entryName}{Name of the entry to be deleted}

\item{verbose}{(logical) default = FALSE}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  The entryName is specified as returned by ls.
}
}
