% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{deleteAllEntries.L2R}
\alias{deleteAllEntries.L2R}
\title{Rsync API}
\usage{
\method{deleteAllEntries}{L2R}(db, verbose = FALSE)
}
\arguments{
\item{db}{RsyncServer object , returning: 'name', 'lastModified' and 'size'}

\item{verbose}{(logical) default = FALSE}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  ...
}
}
