% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{deleteAllEntries.RsyncDHTTP}
\alias{deleteAllEntries.RsyncDHTTP}
\title{Rsync API}
\usage{
\method{deleteAllEntries}{RsyncDHTTP}(db, verbose = FALSE)
}
\arguments{
\item{db}{Rsync object}

\item{verbose}{(logical) default = FALSE}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
 deletes all entries from the Rsync object
}
}
