% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{deleteAllEntries.RsyncL}
\alias{deleteAllEntries.RsyncL}
\title{Rsync API}
\usage{
\method{deleteAllEntries}{RsyncL}(db, verbose = FALSE)
}
\arguments{
\item{db}{RsyncServer object}

\item{verbose}{(logical) default = FALSE}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
 deletes all entries from the destination directory
}
}
