% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteEntry.R
\name{deleteEntry.default}
\alias{deleteEntry.default}
\title{Rsync API}
\usage{
\method{deleteEntry}{default}(db, entryName, verbose = FALSE, ...)
}
\arguments{
\item{db}{rsync object that contains information on the type of connection,
the target directory (remote or local) and eventually a password.}

\item{entryName}{entry that shall be deleted.}

\item{verbose}{FALSE. If set to TRUE, it prints details of the process.}

\item{...}{additional arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
