% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{deleteEntry.RsyncDHTTP}
\alias{deleteEntry.RsyncDHTTP}
\title{Rsync API}
\usage{
\method{deleteEntry}{RsyncDHTTP}(db, entryName, verbose = FALSE)
}
\arguments{
\item{db}{RsyncServer object}

\item{entryName}{Name of the entry to be deleted}

\item{verbose}{(logical) default = FALSE}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  entryName is deleted from the RsyncServer object
}
}
