% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{extractFile.RsyncDHTTP}
\alias{extractFile.RsyncDHTTP}
\title{Rsync API}
\usage{
\method{extractFile}{RsyncDHTTP}(db, file, to, validate = TRUE,
  verbose = FALSE)
}
\arguments{
\item{db}{Rsync object}

\item{file}{file to be sent (i.e .txt file, .R file, etc.)}

\item{validate}{(logical) default: TRUE,}

\item{verbose}{(logical) default: FALSE}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Extracts a file from a db and sends it to a locoal directory.
}
}
