% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{extractFolder}
\alias{extractFolder}
\title{Rsync API}
\usage{
extractFolder(db, ...)
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Extracts a file from a Rsync object.
}
}
