% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{extractFolder.RsyncDHTTP}
\alias{extractFolder.RsyncDHTTP}
\title{Rsync API}
\usage{
\method{extractFolder}{RsyncDHTTP}(db, to, folder = "", ...,
  validate = TRUE, verbose = FALSE)
}
\arguments{
\item{db}{Rsync object}

\item{validate}{(logical) default: TRUE,}

\item{verbose}{(logical) default: FALSE}

\item{file}{file to be sent (i.e .txt file, .R file, etc.)}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Extracts a file from a db and sends it to a locoal directory.
}
}
