% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntry.R
\name{getFile.default}
\alias{getFile.default}
\title{Rsync API}
\usage{
\method{getFile}{default}(db, entryName, validate = TRUE,
  verbose = FALSE, ...)
}
\arguments{
\item{db}{rsync object that contains information on the type of connection,
the target directory (remote or local) and eventually a password.}

\item{entryName}{an entry that the get function will get, located on the
rsync deamon side / the local target directory.}

\item{validate}{validates if entryName is identical in both locations.}

\item{verbose}{FALSE. If set to TRUE, it prints details of the process.}

\item{...}{additional arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
