% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFile.R
\name{getFile}
\alias{getFile}
\title{Rsync API}
\usage{
getFile(host, ...)
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Sends a file to or from a Rsync object.
}
}
