% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{identical}
\alias{identical}
\title{Rsync API}
\usage{
identical(localFile, remoteFile)
}
\arguments{
\item{localFile}{source file}

\item{remoteFile}{target file}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
Tests if two files are exactly identical. It returns TRUE in this case, FALSE in every other case.
}
}
