% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identicalEntries.R
\name{identicalEntries.default}
\alias{identicalEntries.default}
\title{Rsync API}
\usage{
\method{identicalEntries}{default}(db, entryName, ...)
}
\arguments{
\item{db}{rsync object that contains information on the type of connection,
the target directory (remote or local) and eventually a password.}

\item{entryName}{an entry that will be checked}

\item{...}{additional arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
