% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identicalEntries.R
\name{identicalEntries}
\alias{identicalEntries}
\title{Rsync API}
\usage{
identicalEntries(db, ...)
}
\arguments{
\item{db}{rsync object that contains information on the type of connection,
the target directory (remote or local) and eventually a password.}

\item{...}{additional arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
Tests if two files are exactly identical. It returns TRUE in this case,
    FALSE in every other case.
}
}
