% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listEntries.R
\name{listEntries}
\alias{listEntries}
\title{List entries}
\usage{
listEntries(db, ...)
}
\arguments{
\item{db}{rsync object connection}

\item{...}{ignored}
}
\value{
Returns a data.frame.
}
\description{
List all entries in destination folder.
}
