% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{listEntries.RsyncL}
\alias{listEntries.RsyncL}
\title{Rsync API}
\usage{
\method{listEntries}{RsyncL}(db)
}
\arguments{
\item{db}{RsyncServer object}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
Returns a data.frame with the elements in the object.
}
}
