% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFiles.R
\name{listFiles}
\alias{listFiles}
\title{List entries}
\usage{
listFiles(db, ...)
}
\arguments{
\item{db}{rsync object that contains information on the type of connection,
the target directory (remote or local) and eventually a password.}

\item{...}{additional arguments}
}
\description{
List all entries in destination folder. Returns a data frame.
}
\details{
\describe{
  Lists entries of a Rsync object.
}
}
