% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{loadcsv}
\alias{loadcsv}
\title{Rsync API}
\usage{
loadcsv(file, ...)
}
\arguments{
\item{file}{file of type .csv}

\item{...}{more arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
loads a csv file
}
}
