% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{loadjson.default}
\alias{loadjson.default}
\title{Rsync API}
\usage{
\method{loadjson}{default}(db, file, ...)
}
\arguments{
\item{file}{file of type json}

\item{...}{more arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
loads a json file
}
}
