% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadjson.R
\name{loadjson}
\alias{loadjson}
\title{Rsync API}
\usage{
loadjson(host, ...)
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Sends a file to or from a Rsync object.
}
}
