% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{loadrdata}
\alias{loadrdata}
\title{Rsync API}
\usage{
loadrdata(db, ...)
}
\arguments{
\item{address}{from where the object shall be loaded}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  loads a R data object
}
}
