% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{loadrdata.RsyncDHTTP}
\alias{loadrdata.RsyncDHTTP}
\title{Rsync API}
\usage{
\method{loadrdata}{RsyncDHTTP}(db, address)
}
\arguments{
\item{address}{from where the object shall be loaded}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  loads a R data object from a Rsync object
}
}
