% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{ls}
\alias{ls}
\title{Rsync API}
\usage{
ls(db)
}
\arguments{
\item{db}{RsyncServer object , returning: 'name', 'lastModified' and 'size'}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
Returns a data.frame with the elements in the rsync storage.
}
}
