% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{print.RsyncDHTTP}
\alias{print.RsyncDHTTP}
\title{Rsync API}
\usage{
\method{print}{RsyncDHTTP}(x, ...)
}
\arguments{
\item{x}{Rsync object}

\item{...}{more arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
prints the Rsync object
}
}
