% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{print.RsyncServer}
\alias{print.RsyncServer}
\title{Rsync API}
\usage{
\method{print}{RsyncServer}(x, ...)
}
\arguments{
\item{x}{RsyncServer object}

\item{...}{more arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
prints the RsyncServer object
}
}
