% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeAllFiles.R
\name{removeAllFiles.default}
\alias{removeAllFiles.default}
\title{Rsync API}
\usage{
\method{removeAllFiles}{default}(db, verbose = FALSE, ...)
}
\arguments{
\item{db}{rsync object that contains information on the type of connection,
the target directory (remote or local) and eventually a password.}

\item{verbose}{FALSE. If set to TRUE, it prints details of the process.}

\item{...}{additional arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
Deletes all entries from a rsync object.
}
}
