% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeFile.R
\name{removeFile}
\alias{removeFile}
\title{Rsync API}
\usage{
removeFile(db, ...)
}
\arguments{
\item{db}{rsync object that contains information on the type of connection,
the target directory (remote or local) and eventually a password.}

\item{...}{additional arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
Deletes an entry from a Rsync object.
}
}
