% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{newRsync}
\alias{newRsync}
\title{Rsync API}
\usage{
newRsync(from = getwd(), host = NULL, name = NULL,
  to = paste0(host, name), password = NULL)
}
\arguments{
\item{from}{(character) the working directory, if not specified otherwise}

\item{host}{(character) the address to the rsync deamon. NULL for a local rsync connection.}

\item{name}{(character) the name of the deamon's directory. NULL for a local rsync connection.}

\item{to}{(character) a directory, if it is a local rsync connection.}

\item{password}{(character) the password to the corresponding target server.}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
 Establishes a rsync connection with a local directory or a rsync deamon. All arguments are characters.
}
}
