% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntry.R, R/listFiles.R, R/loadData.R,
%   R/removeAllFiles.R, R/removeFile.R, R/rsync.R, R/sendAllFiles.R,
%   R/sendFile.R, R/sendObject.R
\name{getFile}
\alias{getFile}
\alias{getFile.default}
\alias{listFiles}
\alias{listFiles.default}
\alias{loadData}
\alias{loadData.default}
\alias{removeAllFiles}
\alias{removeAllFiles.default}
\alias{removeFile}
\alias{removeFile.default}
\alias{rsync}
\alias{sendAllFiles}
\alias{sendAllFiles.default}
\alias{sendFile}
\alias{sendFile.default}
\alias{sendObject}
\alias{sendObject.default}
\title{\describe{
Deletes an entry.
}}
\usage{
getFile(db, ...)

\method{getFile}{default}(db, fileName, validate = TRUE,
  verbose = FALSE, ...)

listFiles(db, ...)

\method{listFiles}{default}(db, ...)

loadData(db, ...)

\method{loadData}{default}(db, fileName, verbose = FALSE, ...)

removeAllFiles(db, ...)

\method{removeAllFiles}{default}(db, verbose = FALSE, ...)

removeFile(db, ...)

\method{removeFile}{default}(db, fileName, verbose = FALSE, ...)

rsync(dest, src = getwd(), password = NULL)

sendAllFiles(db, ...)

\method{sendAllFiles}{default}(db, validate = FALSE, verbose = FALSE,
  ...)

sendFile(db, ...)

\method{sendFile}{default}(db, fileName, validate = FALSE,
  verbose = FALSE, ...)

sendObject(db, ...)

\method{sendObject}{default}(db, object,
  objectName = as.character(substitute(object)), validate = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{db}{(rsync) an object of class 'rsync' initialized with \code{rsync}.}

\item{...}{arguments passed to methods.}

\item{fileName}{(character) a file name that exists in \code{src}}

\item{validate}{(logical) if the file in dest and src should be validated
using a sha256 check sum.}

\item{verbose}{(logical) if we use 'vorbose' as option in the cli.}

\item{dest}{(character) the address to the rsync daemon or a folder.}

\item{src}{(character) a folder.}

\item{password}{(character|NULL) a password in case a rsync daemon is used.}

\item{object}{(ANY) any R object you wish to store.}

\item{objectName}{(character) the name used to store the object. The file
extension will always be a 'Rdata'.}
}
\description{
Setup a rsync configuration. The configuration object can be used for access
to a folder.
}
\details{
\describe{
  \code{getFile} downloads a file from a Rsync object. If validate is TRUE
  the hash-sum of the remote file is compared to the local version. A warning
  is issued should they differ.
}

\describe{
  \code{loadData}: sync and load the contents of a file. This can be a Rdata
  file or a csv or json.
}

\describe{
  \code{removeAllFiles} deletes all entries from a rsync object.
}

\code{rsync} is a command line tool. For details see
  \url{https://rsync.samba.org/}.

\describe{
  \code{sendAllFiles} Sends all files in src to dest using \code{sendFile}.
}

\describe{
  \code{sendFile} Sends a file to a rsync object.
}

\describe{
\code{sendObject} sends an object (from the environment)
  to a rsync target. This abstracts a pattern where you would use \link{save}
  followed by \code{sendFile}. The reverse is done using \code{getObject}
}
}
