% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsynccli.R
\name{rsynccli}
\alias{rsynccli}
\title{Interface to rsync cli-tool}
\usage{
rsynccli(
  file,
  to,
  includes = NULL,
  excludes = NULL,
  args = "-rltvx",
  pre = NULL,
  intern = FALSE
)
}
\arguments{
\item{file}{(character) source.}

\item{to}{(character) destination.}

\item{includes, excludes}{(character) with length >=1 or (NULL).}

\item{args}{(character) arguments passed to rsync. Default is '-rltvx' and
works for most cases.}

\item{pre}{(character) something that is pasted in front of the 'rsync'
command. E.g. a password.}

\item{intern}{(logical) passed to \link{system}.}
}
\description{
Calls the CLI-program 'rsync'.
}
