% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{rsyncD}
\alias{rsyncD}
\title{Rsync API}
\usage{
rsyncD(host, name, password)
}
\arguments{
\item{host}{(character) the url, where the host file is stored}

\item{name}{(character) the name of the target server}

\item{password}{(character) the password to the corresponding target server}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
 Connection with a Rsync deamon. All arguments are characters.
}
}
