% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{rsyncDHTTP}
\alias{rsyncDHTTP}
\title{Rsync API}
\usage{
rsyncDHTTP(host, name, password, url)
}
\arguments{
\item{host}{(character) the url, where the host file is stored}

\item{name}{(character) the name of the target server}

\item{password}{(character) the password to the corresponding target server}

\item{url}{(character) the url to the corresponding target server}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
 Establishes a connection with a Rsync HTTP server. All arguments are characters.
}
}
