% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsyncFile.R
\name{rsyncFile}
\alias{rsyncFile}
\title{Rsync API}
\usage{
rsyncFile(host, ...)
}
\arguments{
\item{db}{Rsync object}

\item{file}{file to be synced (i.e .txt file, .R file, etc.)}

\item{args}{(character) more arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Syncs a file with db
}
}
