% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{rsyncL}
\alias{rsyncL}
\title{Rsync API}
\usage{
rsyncL(from, to)
}
\arguments{
\item{from}{(character) only for type RsyncL;}

\item{to}{(character) only for type RsyncL}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
 Local connection. All arguments are characters.
}
}
