% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfRsyncServer.R
\docType{data}
\name{RsyncServer}
\alias{RsyncServer}
\title{Rsync API}
\format{An object of class \code{module} (inherits from \code{list}) of length 8.}
\usage{
RsyncServer
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
\item{const(host, name, password, url)}{
  Constructor for a RsyncServer. All arguments are character.
}
\item{ls(db)}{
  Returns a data.frame with the elements in the rsync storage.
}
\item{delete(db, entryName)}{
  The entryName is specified as returned by ls.
}
\item{deleteAllEntries(db)}{
  ...
}
\item{sendFile(db, file, validate = TRUE, verbose = TRUE)}{
  Sends a file (file in local file system) to db. If validate is TRUE the
  hash-sum of the remote file is compared to the local version. A warning is
  issued should the differ. The return status of the command line rsync is
  returned by this function.
}
\item{sendFolder(db, folder, ..., validate = TRUE, verbose = TRUE)}{
  Sends the content of a folder to db using \code{sendFile}. \code{...} are
  passed to \link{dir}. \code{validate} and \code{verbose} are as in
  \code{sendFile}.
}
\item{sendObject(db, obj, ...)}{
  Send an R object to db. This is a wrapper around \code{sendFile}.
  \code{...} are passed to \code{sendFile}.
}
\item{get(db, entryName, file = NULL, ..., loader)}{
  If \code{file} is a character the entry is saved there first. Then it is
  tried to load that file into R and return it from this function. This is
  implemented for csv, json and Rdata files. \code{...} are passed to
  \link{download.file} which is only used in case \code{file} is not
  \code{NULL}.
}
}
}
\keyword{datasets}
