% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendAllFiles.R
\name{sendAllFiles.default}
\alias{sendAllFiles.default}
\title{Rsync API}
\usage{
\method{sendAllFiles}{default}(db, validate = FALSE, verbose = FALSE,
  ...)
}
\arguments{
\item{db}{rsync object that contains information on the type of connection,
the target directory (remote or local) and eventually a password.}

\item{validate}{TRUE. validates if entryName is identical in both locations.}

\item{verbose}{FALSE. If set to TRUE, it prints details of the process.}

\item{...}{additional arguments}

\item{folderName}{folder, of which the content shall be sent}
}
\description{
API to use rsync as persistent file and object storage.
}
