% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendFile.R
\name{sendFile.default}
\alias{sendFile.default}
\title{Rsync API}
\usage{
\method{sendFile}{default}(local, host, fileName, validate = TRUE,
  verbose = FALSE)
}
\arguments{
\item{validate}{(logical) default: TRUE,}

\item{verbose}{(logical) default: FALSE}

\item{db}{Rsync object , returning: 'name', 'lastModified' and 'size'}

\item{file}{file to be sent (i.e .txt file, .R file, etc.)}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Sends a file (file in local file system) to db. If validate is TRUE the
  hash-sum of the remote file is compared to the local version. A warning is
  issued should the differ. The return status of the command line rsync is
  returned by this function.
}
}
