% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{sendFile}
\alias{sendFile}
\title{Rsync API}
\usage{
sendFile(db, ...)
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Sends a file to or from a RsyncServer object.
}
}
