% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{sendFile.RsyncDHTTP}
\alias{sendFile.RsyncDHTTP}
\title{Rsync API}
\usage{
\method{sendFile}{RsyncDHTTP}(db, file, validate = TRUE,
  verbose = FALSE)
}
\arguments{
\item{db}{RsyncServer object}

\item{file}{file to be sent (i.e .txt file, .R file, etc.)}

\item{validate}{(logical) default: TRUE,}

\item{verbose}{(logical) default: FALSE}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Sends a file (file in local file system) to db. If validate is TRUE the
  hash-sum of the remote file is compared to the local version. A warning is
  issued should the differ. The return status of the command line rsync is
  returned by this function.
}
}
