% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{sendFile.RsyncL}
\alias{sendFile.RsyncL}
\title{Rsync API}
\usage{
\method{sendFile}{RsyncL}(db, file, validate = TRUE, verbose = FALSE)
}
\arguments{
\item{db}{RsyncServer object}

\item{validate}{(logical) default: TRUE,}

\item{verbose}{(logical) default: FALSE}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Sends a local file to another local destination directory. If validate is TRUE the
  hash-sum of the remote file is compared to the local version. A warning is
  issued should the differ. The return status of the command line rsync is
  returned by this function.
}
}
