% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendFolder.R
\name{sendFolder}
\alias{sendFolder}
\title{Rsync API}
\usage{
sendFolder(db, ...)
}
\arguments{
\item{db}{rsync object that contains information on the type of connection,
the target directory (remote or local) and eventually a password.}

\item{...}{additional arguments}

\item{folder}{folder, of which the content shall be sent}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Sends the content of a folder to db using \code{sendFile}. \code{validate} and \code{verbose} are as in
  \code{sendFile}.
}
}
