% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{sendObject}
\alias{sendObject}
\title{Rsync API}
\usage{
sendObject(db, obj, objName = as.character(substitute(obj)), ...)
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
\item{sendObject(db, obj, ...)}{
  Send an R object to db. This is a wrapper around \code{sendFile}.
  \code{...} are passed to \code{sendFile}.
}
}
}
