% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{sendObject.RsyncDHTTP}
\alias{sendObject.RsyncDHTTP}
\title{Rsync API}
\usage{
\method{sendObject}{RsyncDHTTP}(db, obj,
  objName = as.character(substitute(obj)), ...)
}
\arguments{
\item{db}{Rsync object , returning: 'name', 'lastModified' and 'size'}

\item{obj}{R object to be sent to db}

\item{objName}{name of 'obj'}

\item{...}{more arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Send an R object to db. This is a wrapper around \code{sendFile}.
  \code{...} are passed to \code{sendFile}.
}
}
