% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awss3.R, R/getFile.R, R/listFiles.R,
%   R/removeFile.R, R/sendAllFiles.R, R/sendFile.R
\name{awss3}
\alias{awss3}
\alias{profileCreate}
\alias{getFile.awss3}
\alias{listFiles.awss3}
\alias{removeFile.awss3}
\alias{sendAllFiles.awss3}
\alias{sendFile.awss3}
\title{Connection object to a AWS S3 bucket}
\usage{
awss3(dest, src = getwd(), profile = NULL)

profileCreate(profile, force = FALSE)

\method{getFile}{awss3}(db, fileName, validate = FALSE, verbose = FALSE, ...)

\method{listFiles}{awss3}(db, ...)

\method{removeFile}{awss3}(db, fileName, verbose = FALSE, ...)

\method{sendAllFiles}{awss3}(db, verbose = FALSE, ...)

\method{sendFile}{awss3}(db, fileName, validate = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{dest, src}{(character) an s3 bucket, e.g. \code{s3://my-bucket} or a
local directory}

\item{profile}{(NULL|character|list) the name of a profile or a list defining
a profile. In case of a list a new profile will be created which is
persistent. A profile is created using \code{aws configure} and stores
credentials for the user in plain text.}

\item{force}{(logical) override profile if it exists.}

\item{db}{(awss3) connection created with \code{awss3}}

\item{fileName}{(character) a file name in dest/src}

\item{validate}{(logical) if validation should take place}

\item{verbose}{(logical) if TRUE print more information to the console}

\item{...}{arguments passed to method}
}
\description{
Only methods specific to this class are documented here. For others the
default method will work. This connection provides the same interface as
\link{rsync}.
}
\examples{
\dontrun{
awss3("s3://my-bucket", profile = list(
  name = "my-profile", # the name of the profile to generate
  aws_access_key_id = "my-access-key-id",
  aws_secret_access_key = "my-secret-access-key",
  region = "my-region"
))
awss3("s3://my-bucket", profile = "my-profile")
}

}
