% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{connection}
\alias{connection}
\title{Rsync API}
\usage{
connection(host = "", name = "", password = "", url = "",
  type = "RsyncDHTTP", from = "", to = "")
}
\arguments{
\item{host}{(character) the url, where the host file is stored}

\item{name}{(character) the name of the target server}

\item{password}{(character) the password to the corresponding target server}

\item{url}{(character) the url to the corresponding target server}

\item{type}{(character) the type of connection. Possible are RsyncDHTTP (default), RsyncD, RsyncL.}

\item{from}{(character) only for type RsyncL;}

\item{to}{(character) only for type RsyncL}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
 Connection for a RsyncServer. All arguments are characters.
}
}
