% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteAllEntries.R
\name{deleteAllEntries}
\alias{deleteAllEntries}
\title{Rsync API}
\usage{
deleteAllEntries(db, ...)
}
\arguments{
\item{db}{rsync object that contains information on the type of connection, the target directory (remote or local) and eventually a password.}

\item{...}{further arguments}

\item{verbose}{FALSE. If set to TRUE, it prints details of the process.}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
Deletes all entries from a rsync object.
}
}
