% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteAllEntries.R
\name{deleteAllEntries}
\alias{deleteAllEntries}
\title{Rsync API}
\usage{
deleteAllEntries(host, ...)
}
\arguments{
\item{db}{Rsync object}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
deletes an entry from a Rsync object
}
}
