% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{deleteAllEntries}
\alias{deleteAllEntries}
\title{Rsync API}
\usage{
deleteAllEntries(db, ...)
}
\arguments{
\item{db}{RsyncServer object}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
 deletes all entries from the RsyncServer object
}
}
