% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{deleteEntry.default}
\alias{deleteEntry.default}
\title{Rsync API}
\usage{
\method{deleteEntry}{default}(db, entryName, verbose = FALSE)
}
\arguments{
\item{db}{Rsync object}

\item{entryName}{Name of the entry to be deleted}

\item{verbose}{(logical) default = FALSE}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  The entryName is specified as returned by ls.
}
}
