% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{deleteEntry}
\alias{deleteEntry}
\title{Rsync API}
\usage{
deleteEntry(db, ...)
}
\arguments{
\item{db}{RsyncServer object}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
deletes an entry from a RsyncServer object
}
}
