% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntry.R
\name{getEntry}
\alias{getEntry}
\title{Rsync API}
\usage{
getEntry(db, ...)
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Sends a file to or from a Rsync object.
}
}
