% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{getEntry}
\alias{getEntry}
\title{Rsync API}
\usage{
getEntry(db, ...)
}
\arguments{
\item{db}{Rsync object}

\item{...}{more arguments}

\item{entryName}{Name of the entry in db}

\item{file}{(character)  default= NULL}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  If \code{file} is a character the entry is saved there first. Then it is
  tried to load that file into R and return it from this function. This is
  implemented for csv, json and Rdata files. \code{...} are passed to
  \link{download.file} which is only used in case \code{file} is not
  \code{NULL}.
}
}
