% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{listDir}
\alias{listDir}
\title{Rsync API}
\usage{
listDir(dirName)
}
\arguments{
\item{x}{Rsync object}

\item{...}{more arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
converts a RsynsServer object to a character
}
}
