% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{listEntries.default}
\alias{listEntries.default}
\title{Rsync API}
\usage{
\method{listEntries}{default}(db)
}
\arguments{
\item{db}{Rsync object}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
Returns a data.frame with the elements in the rsync object.
}
}
