% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{listEntries}
\alias{listEntries}
\title{Rsync API}
\usage{
listEntries(db, ...)
}
\arguments{
\item{host}{(character) the url, where the host file is stored}

\item{name}{(character) the name of the target server}

\item{password}{(character) the password to the corresponding target server}

\item{url}{(character) the url to the corresponding target server}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
 Connection for a RsyncServer. All arguments are characters.
}


#' @export
connectionOld <- function(host, name, password, url) {
  # validate input here
  host <- paste0(sub("/$", "", host), "/")
  name <- sub("^/", "", name)
  url <- paste0(sub("/$", "", url), "/")
  aoos::retList("RsyncServer")
}
#' @export
connectionL2R <- function(host, name, password, url) {
  # validate input here
  host <- paste0(sub("/$", "", host), "/")
  name <- sub("^/", "", name)
  url <- paste0(sub("/$", "", url), "/")
  aoos::retList("L2R")
}
}
