% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listEntries.R
\name{listEntries}
\alias{listEntries}
\title{List entries}
\usage{
listEntries(host, ...)
}
\arguments{
\item{host}{rsync object connection}

\item{...}{ignored}
}
\description{
List all entries in destination folder. Returns a data frame.
}
