% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{listEntries}
\alias{listEntries}
\title{Rsync API}
\usage{
listEntries(db, ...)
}
\arguments{
\item{db}{a Rsync object that stores necessary information for respective connection types}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
 Lists entries of db objects.
}
}
