% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listEntries.R
\name{listEntries}
\alias{listEntries}
\title{List entries}
\usage{
listEntries(db, ...)
}
\arguments{
\item{db}{rsync object that contains information on the type of connection, the target directory (remote or local) and eventually a password.}

\item{...}{further arguments}
}
\description{
List all entries in destination folder. Returns a data frame.
}
\details{
\describe{
  Lists entries of a Rsync object.
}
}
