% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadcsv.R
\name{loadcsv}
\alias{loadcsv}
\title{Rsync API}
\usage{
loadcsv(host, ...)
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Sends a file to or from a Rsync object.
}
}
