% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadcsv.R
\name{loadcsv}
\alias{loadcsv}
\title{Rsync API}
\usage{
loadcsv(host, ...)
}
\arguments{
\item{host}{Rsync object, with csv file}

\item{...}{}

\item{csvName}{name of csv file}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  loads a csv file from a Rsync object.
}
}
