% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadjson.R
\name{loadjson}
\alias{loadjson}
\title{Rsync API}
\usage{
loadjson(db, ...)
}
\arguments{
\item{...}{}

\item{host}{Rsync object, with json file}

\item{jsonName}{name of json file}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Loads a file from a Rsync object.
}
}
