% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadjson.R
\name{loadjson}
\alias{loadjson}
\title{Rsync API}
\usage{
loadjson(db, ...)
}
\arguments{
\item{db}{rsync object that contains information on the type of connection, the target directory (remote or local) and eventually a password.}

\item{...}{additional arguments}

\item{jsonName}{name of json file}

\item{verbose}{FALSE. If set to TRUE, it prints details of the process.}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Loads a file from a Rsync object.
}
}
