% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{loadjson}
\alias{loadjson}
\title{Rsync API}
\usage{
loadjson(db, ...)
}
\arguments{
\item{...}{more arguments}

\item{file}{file of type json}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
loads a json file
}

#' @export
}
