% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadrdata.R
\name{loadrdata}
\alias{loadrdata}
\title{Rsync API}
\usage{
loadrdata(db, ...)
}
\arguments{
\item{...}{}

\item{host}{Rsync object, with csv file}

\item{csvName}{name of csv file}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  loads a rdata file from a Rsync object.
}
}
