% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsync.R
\name{rsync}
\alias{rsync}
\title{Rsync API}
\usage{
rsync(dest, src = getwd(), password = NULL)
}
\arguments{
\item{dest}{(character) the address to the rsync deamon. NULL for a
src rsync connection.}

\item{src}{(character) the working directory, if not specified otherwise}

\item{password}{(character|NULL) the password to the corresponding target
server.}
}
\description{
Setup a rsync configuration. The configuration object can be used for remote
or src access to a folder. For available methods, see the 'seeAlso' section
of this documentation.
}
\details{
\code{rsync} is a command line tool. For details see
  \url{https://rsync.samba.org/}.
}
\seealso{
\link{getFile}, \link{removeFile}, \link{listFiles},
  \link{sendFile}, \link{sendFolder}, \link{sendObject}
}
