% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsync.R
\name{rsync}
\alias{rsync}
\title{Interface to rsync cli-tool}
\usage{
rsync(from, to, includes = NULL, excludes = NULL, args = "-rltvx",
  pre = NULL)
}
\arguments{
\item{from}{(character) source}

\item{to}{(character) target}

\item{includes, excludes}{(chracter) with length >=1 restictions}

\item{args}{(character) more arguments}

\item{pre}{(character) something which is pasted before the 'rsync' command}

\item{...}{arguments passed to rsync}
}
\description{
Calls the CLI-program 'rsync'.
}
