% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsyncFile.R
\name{rsyncFile}
\alias{rsyncFile}
\title{Rsync API}
\usage{
rsyncFile(host, ...)
}
\arguments{
\item{host}{Rsync object}

\item{local}{local directory}

\item{fileName}{file to be synced (i.e .txt file, .R file, etc.)}

\item{direction}{the direction that it is synced in ('send'- from local to host or 'get' - from host to local)}

\item{args}{(character) more arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Syncs a file with of a rsync object
}
}
