% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{rsyncFile}
\alias{rsyncFile}
\title{Rsync API}
\usage{
rsyncFile(db, file, args)
}
\arguments{
\item{db}{RsyncServer object , returning: 'name', 'lastModified' and 'size'}

\item{file}{file to be synced (i.e .txt file, .R file, etc.)}

\item{args}{(character) more arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Syncs a file to a db
}
}
