% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsyncSuccessful.R
\name{rsyncSuccessful}
\alias{rsyncSuccessful}
\title{Rsync API}
\usage{
rsyncSuccessful(local, host, filenName)
}
\arguments{
\item{localFile}{source file}

\item{remoteFile}{target file}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
Tests if the sync process was successful. It returns TRUE in this case, FALSE in every other case.
}
}
