% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendFile.R
\name{sendFile.default}
\alias{sendFile.default}
\title{Rsync API}
\usage{
\method{sendFile}{default}(db, fileName, validate = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{db}{rsync object that contains information on the type of connection,
the target directory (remote or local) and eventually a password.}

\item{fileName}{file, which shall be sent to a target directory}

\item{validate}{TRUE. validates if entryName is identical in both locations.}

\item{verbose}{FALSE. If set to TRUE, it prints details of the process.}

\item{...}{additional arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
