% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendFile.R
\name{sendFile.default}
\alias{sendFile.default}
\title{Rsync API}
\usage{
\method{sendFile}{default}(host, local, fileName, verbose = FALSE)
}
\arguments{
\item{host}{Rsync object , returning: 'name', 'lastModified' and 'size'}

\item{local}{local directory to send file from}

\item{fileName}{file to be sent (i.e .txt file, .R file, etc.)}

\item{verbose}{(logical) default: FALSE}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Sends a file (file in local file system) to a host.
}
}
