% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendFile.R
\name{sendFile}
\alias{sendFile}
\title{Rsync API}
\usage{
sendFile(host, ...)
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Sends a file to a rsync host object.
}
}
