% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{sendFolder.default}
\alias{sendFolder.default}
\title{Rsync API}
\usage{
\method{sendFolder}{default}(db, to, ..., validate = TRUE,
  verbose = FALSE)
}
\arguments{
\item{db}{RsyncServer object , returning: 'name', 'lastModified' and 'size'}

\item{...}{more arguments}

\item{validate}{(logical) default = TRUE}

\item{verbose}{(logical) default = TRUE}

\item{folder}{folder, of which the content shall be sent}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Sends the content of a folder to db using \code{sendFile}. \code{...} are
  passed to \link{dir}. \code{validate} and \code{verbose} are as in
  \code{sendFile}.
}
}
