% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendFolder.R
\name{sendFolder.default}
\alias{sendFolder.default}
\title{Rsync API}
\usage{
\method{sendFolder}{default}(db, folderName, validate = TRUE,
  verbose = FALSE, ...)
}
\arguments{
\item{db}{rsync object that contains information on the type of connection,
the target directory (remote or local) and eventually a password.}

\item{folderName}{folder, of which the content shall be sent}

\item{validate}{TRUE. validates if entryName is identical in both locations.}

\item{verbose}{FALSE. If set to TRUE, it prints details of the process.}

\item{...}{additional arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
