% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{sendFolder}
\alias{sendFolder}
\title{Rsync API}
\usage{
sendFolder(db, folder, ..., validate = TRUE, verbose = FALSE)
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
\item{sendFolder(db, folder, ..., validate = TRUE, verbose = TRUE)}{
  Sends the content of a folder to db using \code{sendFile}. \code{...} are
  passed to \link{dir}. \code{validate} and \code{verbose} are as in
  \code{sendFile}.
}
}
}
