% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendFolder.R
\name{sendFolder}
\alias{sendFolder}
\title{Rsync API}
\usage{
sendFolder(host, ...)
}
\arguments{
\item{...}{more arguments}

\item{db}{Rsync object , returning: 'name', 'lastModified' and 'size'}

\item{folder}{folder, of which the content shall be sent}

\item{validate}{(logical) default = TRUE}

\item{verbose}{(logical) default = TRUE}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Sends the content of a folder to db using \code{sendFile}. \code{...} are
  passed to \link{dir}. \code{validate} and \code{verbose} are as in
  \code{sendFile}.
}
}
