% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendObject.R
\name{sendObject.default}
\alias{sendObject.default}
\title{Rsync API}
\usage{
\method{sendObject}{default}(db, object,
  objectName = as.character(substitute(object)), validate = TRUE,
  verbose = FALSE, ...)
}
\arguments{
\item{db}{rsync object that contains information on the type of connection,
the target directory (remote or local) and eventually a password.}

\item{object}{object in the environment, which shall be sent to a target
directory}

\item{objectName}{the object's name that is constructed from `object`}

\item{validate}{TRUE. validates if entryName is identical in both locations.}

\item{verbose}{FALSE. If set to TRUE, it prints details of the process.}

\item{...}{additional arguments}
}
\description{
API to use rsync as persistent file and object storage.
}
