% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendObject.R
\name{sendObject.default}
\alias{sendObject.default}
\title{Rsync API}
\usage{
\method{sendObject}{default}(host, object,
  objectName = as.character(substitute(object)), verbose = FALSE)
}
\arguments{
\item{verbose}{(logical) default: FALSE}

\item{db}{Rsync object , returning: 'name', 'lastModified' and 'size'}

\item{file}{file to be sent (i.e .txt file, .R file, etc.)}

\item{validate}{(logical) default: TRUE,}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Sends a file (file in local file system) to db. If validate is TRUE the
  hash-sum of the remote file is compared to the local version. A warning is
  issued should the differ. The return status of the command line rsync is
  returned by this function.
}
}
