% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsyncServer.R
\name{sendObject}
\alias{sendObject}
\title{Rsync API}
\usage{
sendObject(db, ...)
}
\arguments{
\item{db}{RsyncServer object}

\item{...}{more arguments}

\item{obj}{R object to be sent to db}

\item{objName}{name of 'obj'}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Send an R object to db. This is a wrapper around \code{sendFile}.
  \code{...} are passed to \code{sendFile}.
}
}
