% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendObject.R
\name{sendObject}
\alias{sendObject}
\title{Rsync API}
\usage{
sendObject(host, ...)
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Sends a file to or from a Rsync object.
}
}
