% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendObject.R
\name{sendObject}
\alias{sendObject}
\title{Rsync API}
\usage{
sendObject(db, ...)
}
\arguments{
\item{db}{rsync object that contains information on the type of connection, the target directory (remote or local) and eventually a password.}

\item{...}{additional arguments}

\item{object}{object in the environment, which shall be sent to a target directory}

\item{validate}{TRUE. validates if entryName is identical in both locations.}

\item{verbose}{FALSE. If set to TRUE, it prints details of the process.}
}
\description{
API to use rsync as persistent file and object storage.
}
\details{
\describe{
  Sends an object (from the environment) to a rsync target. If validate is TRUE the
  hash-sum of the remote file is compared to the local version. A warning is
  issued should they differ.
}
}
